requirements.txt
django
django-allauth
PyJWT
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
django-config-models
